requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
matplotlib==3.7.1
tqdm==4.65.0
scikit-image==0.21.0
requests==2.28.2
Pillow==9.5.0
opencv-python==4.7.0.72
scipy==1.10.1